Make clippy a little bit more happy
