+++ aquadoggo/src/replication/strategies/mod.rs
@@ -1,8 +1,10 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+mod diff;
 mod naive;
 mod set_reconciliation;
 
+pub use diff::diff_log_heights;
 pub use naive::NaiveStrategy;
 pub use set_reconciliation::SetReconciliationStrategy;
